pyproject.toml
[project]
name = "reco_service"
version = "0.0.1"
description = ""
authors = []
maintainers = []
readme = "README.md"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "implicit (>=0.7.2,<0.8.0)",
    "numpy (>=1.22,<2.0.0)",
    "orjson (>=3.10.15,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "rectools[lightfm,nmslib] (>=0.12.0,<0.13.0)",
    "requests (>=2.32.3,<3.0.0)",
    "starlette (>=0.46.1,<0.47.0)",
    "uvicorn (>=0.34.0,<0.35.0)"
]
requires-python = ">=3.11,<3.13"

[tool.poetry]
packages = [{ include = "service" }]

[tool.poetry.group.dev.dependencies]
bandit = "~1.8.3"
black = "~25.1.0"
flake8 = "~7.1.2"
isort = "~6.0.1"
mypy = "~1.15.0"
pylint = "~3.3.4"
pytest = "~8.3.5"
requests = "~2.32.3"

[tool.black]
line-length = 120
target-version = ["py311", "py312"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
